Revert "Sửa chức năng quản lý đất"
This reverts commit 27a29e20050a93eaf9af4c411fb057bb53fbeab8.

+++ CGP.Domain/Entities/Voucher.cs
@@ -17,7 +17,7 @@ namespace CGP.Domain.Entities
         public PaymentMethodEnum PaymentMethod { get; set; }
         public double MinOrderValue { get; set; }
         public double MaxDiscountAmount { get; set; }
-        public int PointChangeAmount { get; set; }
+        public double PointChangeAmount { get; set; }
         public int UsedCount { get; set; } = 0;
         public int ChangeAmout { get; set; } = 0;
         public int Quantity { get; set; }
